MarkdownParsing: fix issue where TaskItems not always found
Issue was with parsing files that did not have a line ending at the end.
This was fine if the last line in the file was empty or if it was a task
item, but it failed to parse if it was a non-task line that contained
at least one character.

Fix was to append a "\n" to the string before parsing.  Would be better
to work out how to make the parser tollerant to this particular case
(have added to the TODO)